% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbt_write_quarto_bib.R
\name{bbt_write_quarto_bib}
\alias{bbt_write_quarto_bib}
\title{Scan citations and write a bibliography file for a whole R project}
\usage{
bbt_write_quarto_bib(
  bib_file = here::here("references.json"),
  dir = c("", "qmd", "tex"),
  pattern = "\\\\.qmd$|\\\\.tex$",
  ignore = NULL,
  translator = rbbt::bbt_guess_translator(bib_file),
  library_id = getOption("rbbt.default.library_id", 1),
  overwrite = TRUE,
  filter = identity,
  wd = here::here()
)
}
\arguments{
\item{bib_file}{(Optional) A \code{\link[base:character]{character}} string with the
path to the bibliography file (Default: \code{here::here("references.json")}).}

\item{dir}{(Optional) A \code{\link[base:character]{character}} vector indicating the
directories to scan for citation keys (Default: \code{c("", "qmd", "tex")}).}

\item{pattern}{(Optional) A \code{\link[base:character]{character}} string with a
regular expression to match files to scan for citation keys
(Default: \code{"\\.qmd$|\\.tex$"}).}

\item{ignore}{(Optional) A \code{\link[base:character]{character}} string with a
regular expression to match files to ignore when scanning for citation
keys (Default: \code{NULL}).}

\item{translator}{(Optional) A \code{\link[base:character]{character}} string with
the name of the translator to use when writing the bibliography file.
The available translators are: \code{json} (CSL-JSON), \code{biblatex}, \code{bibtex}
(BibLaTeX), and \code{yaml} (CSL YAML). CSL-JSON is recommended if users are
not specifically using a LaTeX citation processor. See
\code{\link[rbbt:bbt_write_bib]{bbt_write_bib()}} for more details
(Default: \code{rbbt::bbt_guess_translator(bib_file)}).}

\item{library_id}{(Optional) A \code{\link[base:numeric]{numeric}} integer with the
Zotero library ID to use when writing the bibliography file.
See \code{\link[rbbt:bbt_write_bib]{bbt_write_bib()}} for more details
(Default: \code{getOption("rbbt.default.library_id", 1)}).}

\item{overwrite}{(Optional) A \code{\link[base:logical]{logical}} flag indicating
whether to overwrite the bibliography file if it already exists
(Default: \code{TRUE}).}

\item{filter}{(Optional) A \code{\link[base:function]{function}} to filter the
bibliography entries before writing the bibliography file. See
\code{\link[rbbt:bbt_write_bib]{bbt_write_bib()}} for more details
(Default: \code{identity}).}

\item{wd}{(Optional) A \code{\link[base:character]{character}} string with the path
to the working directory (Default: \code{here::here()}).}
}
\value{
An invisible \code{NULL}. This function is used for its side effect.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{bbt_write_quarto_bib()} write a bibliography file scanning all citation
keys in a R project.

Different from \code{\link[rbbt:bbt_write_bib]{bbt_write_bib()}}, this function
scans all files in a project directory and subdirectories for citation
keys, ensuring that all citations are included in the bibliography file.
}
\details{
\subsection{Better BibTeX}{

To use this function, you must have the BetterBibTeX
(\href{https://retorque.re/zotero-better-bibtex/}{BBT}) extension installed in
your Zotero. Zotero must be open and running when you execute this function.
}

\subsection{Note}{

(2025-03-10) Unfortunately, the author of the \code{rbbt} package has not yet
merged a
\href{https://github.com/paleolimbot/rbbt/pull/48}{crucial pull request} into
the package repository. If you encounter problems with this function,
please try downloading and installing this \code{rbbt} fork:
\url{https://github.com/wmoldham/rbbt}. You can install it using the following
code:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{remotes::install_github("wmoldham/rbbt")
}\if{html}{\out{</div>}}
}
}
\examples{
\dontrun{
  bbt_write_quarto_bib(
    bib_file = "references.bib",
    dir = c("", "qmd", "tex"),
    pattern = c("\\\\.qmd$|\\\\.tex$"),
    wd = here::here()
 )
}
}
\concept{Zotero functions}
