% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quartor-package.R
\docType{package}
\name{quartor-package}
\alias{quartor}
\alias{quartor-package}
\title{quartor: Quarto Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools designed to facilitate the use of the Quarto publishing system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danielvartan.github.io/quartor/}
  \item \url{https://github.com/danielvartan/quartor/}
  \item Report bugs at \url{https://github.com/danielvartan/quartor/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

}
\keyword{internal}
